client/requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
python-dotenv==1.0.1
python-multipart==0.0.18
requests==2.32.3
RPi.GPIO==0.7.1
typing_extensions==4.12.2
urllib3==2.2.3
